refactor: Externalize spark session creation to utils module
